+++ src/components/base/index.ts
@@ -1,3 +1,4 @@
+export { default as IconButton } from "./IconButton";
 export { default as Button } from "./Button";
 export { default as Tag } from "./Tag";
 export { default as Text } from "./Text";